+++ CHANGELOG.md
@@ -226,6 +226,9 @@ Thanks to the people who made this release happen!
 
 * When reconfiguring the author, warn that the working copy won't be updated
 
+* `jj rebase -s` can now be used with the `--insert-after` and `--insert-before`
+  options, like `jj rebase -r`.
+
 ### Fixed bugs
 
 * Release binaries for Intel Macs have been restored. They were previously